+++ internal/pg_documentdb_distributed/src/test/regress/sql/bson_aggregation_object_operators_tests.sql
@@ -119,10 +119,15 @@ SELECT * FROM bson_dollar_project('{"a": 1}', '{ "result": { "$bsonSize": { "a":
 SELECT * FROM bson_dollar_project('{ "_id": 16, "group": 2, "obj": { "a": 1, "b": 1 } }', '{ "mergedDocument": { "$mergeObjects": ["$obj", { "b": "$$REMOVE" } ] } }');
 
 -- $getField operator
+SELECT insert_one('db','test_get_field',' { "_id": 0, "a": 1, "b": "test" }');
 -- positive cases
 -- entire expression
 -- field parsed from $literal
 SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": {"field": { "$literal": "a" }, "input": {"a": { "b": 3 }}}}}}');
+-- field parsed from expression
+SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": {"field": { "$literal": "a" }, "input": {"a": { "b": 3 }}}}}}');
+-- field is a path
+SELECT document FROM bson_aggregation_pipeline('db', '{ "aggregate": "test_get_field", "pipeline":  [{"$project": {"result": {"fieldValue": {"$getField": {"field": "$b", "input": {"test": { "b": 3 }}}}}}}]}');
 -- input be a system variable
 SELECT * FROM bson_dollar_project('{"a": { "b": 3 }}', '{"result": { "fieldValue": {"$getField": {"field": "a", "input": "$$ROOT"}}}}');
 -- input be null
@@ -164,13 +169,9 @@ SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField"
 SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": {"field": null, "input": {}}}}}');
 SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": {"field": 1, "input": {}}}}}');
 SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": {"field": [], "input": {}}}}}');
--- field is a path
-SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": {"field": "$a", "input": {"a": { "b": 3 }}}}}}');
 -- shorthand expression
 -- field must be a string
 SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": []}}}');
--- field is an operator
-SELECT * FROM bson_dollar_project('{}', '{"result": { "fieldValue": {"$getField": { "$add": [2, 3 ]}}}}');
 
 -- $unsetField
 -- postive cases